[CHORE] User 패키지로 수정
