+++ data/src/main/java/com/nexters/knownknowns/data/repository/NewsRepository.kt
@@ -1,13 +1,8 @@
 package com.nexters.knownknowns.data.repository
 
-import com.nexters.knownknowns.data.local.ClickState
 import com.nexters.knownknowns.data.model.NewsResponse
 import kotlinx.coroutines.flow.Flow
 
 interface NewsRepository {
     fun getNews(): Flow<List<NewsResponse>>
-    suspend fun incrementClickCount()
-    suspend fun resetClickState()
-    suspend fun recordBottomSheetShown()
-    val clickStateFlow: Flow<ClickState>
 }
